+++ crates/subspace-service/src/lib.rs
@@ -31,8 +31,8 @@ pub mod dsn;
 mod metrics;
 pub(crate) mod mmr;
 pub mod rpc;
-mod spawn_tasks;
 pub mod sync_from_dsn;
+mod task_spawner;
 pub mod transaction_pool;
 
 use crate::config::{ChainSyncMode, SubspaceConfiguration, SubspaceNetworking};
@@ -1165,7 +1165,7 @@ where
     // We replace the Substrate implementation of metrics server with our own.
     config.base.prometheus_config.take();
 
-    let rpc_handlers = spawn_tasks::spawn_tasks(SpawnTasksParams {
+    let rpc_handlers = task_spawner::spawn_tasks(SpawnTasksParams {
         network: network_service.clone(),
         client: client.clone(),
         keystore: keystore_container.keystore(),
